syntax: fix invalid parsing with square brackets in desc
Make xitCheckbox*Desc end with either "[" or title *immediately* after
the start of a line, instead of ending whenever it finds a "[".

Use negative lookahead in xitInvalid regex so that a line with a valid
checkbox that consists of \[[x~@ ]\] is immediately treated as valid.
Also only check for the checkbox at the start of the line, so characters
enclosed by square brackets in the middle of a description don't get
considered invalid.